+++ src/service.cpp
@@ -11,7 +11,6 @@
 
 #include <iostream>
 #include <iomanip>
-#include <sstream>
 #include <fstream>
 #include <random>
 
@@ -23,9 +22,9 @@
 
 namespace ecl {
 
-bool ScalerService::keep_running = true;
+bool Service::keep_running = true;
 
-ScalerService::ScalerService(const ServiceOption &option) noexcept
+Service::Service(const ServiceOption &option) noexcept
 : port_(option.port)
 , log_level_(option.log_level)
 , test_(option.test)
@@ -122,7 +121,7 @@ ScalerService::ScalerService(const ServiceOption &option) noexcept
 }
 
 
-ScalerService::~ScalerService() {
+Service::~Service() {
 	if (test_) {
 		if (memory_) delete memory_;
 	} else {
@@ -142,11 +141,11 @@ ScalerService::~ScalerService() {
 
 
 void SigIntHandler(int) {
-	ScalerService::keep_running = false;
+	Service::keep_running = false;
 }
 
 
-void ScalerService::PrintScaler() const noexcept {
+void Service::PrintScaler() const noexcept {
 	signal(SIGINT, SigIntHandler);
 	if (system("tput smcup")) {
 		std::cout << "[Error] Use bash command tput smcup failed.\n";
@@ -158,8 +157,8 @@ void ScalerService::PrintScaler() const noexcept {
 			exit(-1);
 		}
 		std::cout << "scaler      counts\n";
-		for (size_t i = 0; i < kMaxScalers; ++i) {
-			printf("%2ld%15d\n", i, memory_->scaler[i].value);
+		for (uint32_t i = 0; i < kMaxScalers; ++i) {
+			printf("%2d%15d\n", i, memory_->scaler[i].value);
 		}
 		usleep(1000000);
 	}
@@ -182,7 +181,7 @@ std::string GetFileName(
 	std::string device_name,
 	tm *t
 ) {
-	std::stringstream file_name;
+	std::stringstream file_name("");
 	file_name << data_path << t->tm_year+1900
 		<< std::setw(2) << std::setfill('0') << t->tm_mon+1
 		<< std::setw(2) << std::setfill('0') << t->tm_mday
@@ -192,7 +191,7 @@ std::string GetFileName(
 }
 
 
-int ScalerService::ReadDateScaler(
+int Service::ReadDateScaler(
 	tm* date,
 	int32_t flag,
 	size_t seconds,
@@ -252,7 +251,7 @@ int ScalerService::ReadDateScaler(
 }
 
 
-int ScalerService::ReadRecentScaler(
+int Service::ReadRecentScaler(
 	int32_t flag,
 	int seconds,
 	int average,
@@ -405,7 +404,7 @@ int GetFileStream(const char *file_name, std::fstream &fout) {
 }
 
 
-int ScalerService::WriteScaler() const noexcept {
+int Service::WriteScaler() const noexcept {
 	// generate file name according to the date
 	time_t current_time = time(NULL);
 	tm *current_tm = localtime(&current_time);
@@ -439,7 +438,7 @@ int ScalerService::WriteScaler() const noexcept {
 }
 
 
-void ScalerService::Serve() noexcept {
+void Service::Serve() noexcept {
 	// server address
 	std::string server_address = "0.0.0.0:" + std::to_string(port_);
 	// server builder
@@ -461,7 +460,7 @@ void ScalerService::Serve() noexcept {
 }
 
 
-grpc::ServerUnaryReactor* ScalerService::GetState(
+grpc::ServerUnaryReactor* Service::GetState(
 	grpc::CallbackServerContext* context,
 	const Request*,
 	Response *response
@@ -473,7 +472,7 @@ grpc::ServerUnaryReactor* ScalerService::GetState(
 }
 
 
-grpc::ServerWriteReactor<Response>* ScalerService::GetScaler(
+grpc::ServerWriteReactor<Response>* Service::GetScaler(
 	grpc::CallbackServerContext*,
 	const Request*
 ) {
@@ -568,7 +567,7 @@ private:
 };
 
 
-grpc::ServerWriteReactor<Response>* ScalerService::GetScalerRecent(
+grpc::ServerWriteReactor<Response>* Service::GetScalerRecent(
 	grpc::CallbackServerContext*,
 	const RecentRequest* request
 ) {
@@ -611,7 +610,7 @@ grpc::ServerWriteReactor<Response>* ScalerService::GetScalerRecent(
 }
 
 
-grpc::ServerWriteReactor<Response>* ScalerService::GetScalerDate(
+grpc::ServerWriteReactor<Response>* Service::GetScalerDate(
 	grpc::CallbackServerContext*,
 	const DateRequest *request
 ) {
@@ -645,7 +644,7 @@ grpc::ServerWriteReactor<Response>* ScalerService::GetScalerDate(
 }
 
 
-grpc::ServerWriteReactor<Expression>* ScalerService::GetConfig(
+grpc::ServerWriteReactor<Expression>* Service::GetConfig(
 	grpc::CallbackServerContext*,
 	const Request*
 ) {
@@ -681,6 +680,10 @@ grpc::ServerWriteReactor<Expression>* ScalerService::GetConfig(
 		std::vector<Expression> expressions_;
 		size_t index_;
 	};
+
+	if (log_level_ >= kDebug) {
+		std::cout << "[Debug] GetConfig().\n";
+	}
 
 	// expressions
 	std::vector<Expression> expressions;
@@ -733,22 +736,25 @@ grpc::ServerWriteReactor<Expression>* ScalerService::GetConfig(
 }
 
 
-grpc::ServerReadReactor<Expression>* ScalerService::SetConfig(
+grpc::ServerReadReactor<Expression>* Service::SetConfig(
 	grpc::CallbackServerContext*,
-	Response *response
+	ParseResponse *response
 ) {
 	class Recorder : public grpc::ServerReadReactor<Expression> {
 	public:
 		Recorder(
-			Response *response,
+			ParseResponse *response,
+			volatile Memory* memory,
 			bool test,
 			LogLevel log_level
-		): response_(response), test_(test), log_level_(log_level) {
+		): response_(response), memory_(memory), test_(test), log_level_(log_level) {
 			// initialize
 			response_->set_value(0);
 			if (log_level_ >= kDebug) {
 				std::cout << "[Debug] Start to read config.\n";
 			}
+			success_ = true;
+			index_ = 0;
 			StartRead(&expression_);
 		}
 
@@ -758,66 +764,29 @@ grpc::ServerReadReactor<Expression>* ScalerService::SetConfig(
 					std::cout << "[Info] Read expression from client "
 						<< expression_.value() << "\n";
 				}
-				if (config_parser_.Parse(expression_.value())) {
-					response_->set_value(-1);
-					Finish(grpc::Status(
-						grpc::StatusCode::INVALID_ARGUMENT,
-						"Parse expression failed"
-					));
+				ParseResult parse_result =
+					config_parser_.Parse(expression_.value());
+				if (!parse_result.Ok()) {
+					success_ = false;
+					response_->set_value(parse_result.Status());
+					response_->set_index(index_);
+					response_->set_position(int(parse_result.Position()));
+					response_->set_length(int(parse_result.Length()));
+					Finish(grpc::Status::OK);
 				}
+				++index_;
 				StartRead(&expression_);
 			} else {
+				if (!success_) return;
+
 				if (log_level_ >= kDebug) {
 					std::cout << "[Debug] Read expressions done.\n";
 				}
 				// read config from parser
 				memory_config_.Read(&config_parser_);
 				if (!test_) {
-					grpc::Status bad = grpc::Status(
-						grpc::StatusCode::UNAVAILABLE, "mmap failed"
-					);
-					// open memory file
-					int fd = open("/dev/uio0", O_RDWR);
-					if (fd < 0) {
-						if (log_level_ >= kDebug) {
-							std::cout << "[Debug] Failed to open /dev/uio0.\n";
-						}
-						response_->set_value(-2);
-						Finish(bad);
-					}
-					// lock the address space
-					if (flock(fd, LOCK_EX | LOCK_NB)) {
-						if (log_level_ >= kDebug) {
-							std::cout << "[Debug] Failed to acquire file lock.\n";
-						}
-						response_->set_value(-2);
-						Finish(bad);
-					}
-					// map memory
-					void *map_addr = mmap(
-						NULL,
-						4096,
-						PROT_READ | PROT_WRITE,
-						MAP_SHARED,
-						fd,
-						0
-					);
-					if (map_addr == MAP_FAILED) {
-						if (log_level_ >= kDebug) {
-							std::cout << "[Debug] Failed to map.\n";
-						}
-						response_->set_value(-2);
-						Finish(bad);
-					}
-					volatile uint32_t *map = (uint32_t*)map_addr;
-
 					// write config to memory
-					memory_config_.MapMemory(map);
-
-					// clean up
-					flock(fd, LOCK_UN);
-					munmap(map_addr, 4096);
-					close(fd);
+					memory_config_.MapMemory((uint32_t*)memory_);
 				}
 
 				// save backup
@@ -837,15 +806,22 @@ grpc::ServerReadReactor<Expression>* ScalerService::SetConfig(
 		}
 
 	private:
-		Response *response_;
+		ParseResponse *response_;
+		volatile Memory *memory_;
 		bool test_;
 		LogLevel log_level_;
 		Expression expression_;
 		MemoryConfig memory_config_;
 		ConfigParser config_parser_;
+		bool success_;
+		int index_;
 	};
 
-	return new Recorder(response, test_, log_level_);
+	if (log_level_ >= kDebug) {
+		std::cout << "[Debug] SetConfig().\n";
+	}
+
+	return new Recorder(response, memory_, test_, log_level_);
 }
 
 }