+++ test/standardize/test_standard_logic_downscale_tree.cpp
@@ -75,12 +75,12 @@ TEST(StandardLogicDownscaleTreeTest, Output) {
 	for (size_t i = 0; i < kOutput.size(); ++i) {
 		Lexer lexer;
 		LogicDownscaleGrammar grammar;
-		SLRSyntaxParser parser(&grammar);
+		SLRSyntaxParser<int> parser(&grammar);
 		std::vector<TokenPtr> tokens;
 
-		ASSERT_EQ(lexer.Analyse("LEFT="+kExpression[i], tokens), 0)
+		ASSERT_TRUE(lexer.Analyse("LEFT="+kExpression[i], tokens).Ok())
 			<< "Error: Lexer analyse " << i;
-		ASSERT_EQ(parser.Parse(tokens), 0)
+		ASSERT_TRUE(parser.Parse(tokens).Ok())
 			<< "Error: Parser parse " << i;
 
 		ASSERT_EQ(parser.Root()->Eval(), kLayer[i])